\name{pgmm}
\alias{pgmm}
\docType{package}
\title{Parsimonious Gaussian Mixture Models}
\description{
An implementation of model-based clustering and model-based classification using parsimonious Gaussian mixture models.
}
\details{
\tabular{ll}{
Package: \tab pgmm\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2014-07-12\cr
License: \tab GPL (>=2)\cr
}
This package contains the function \code{\link{pgmmEM}} plus three data sets.
}
\author{
Paul D. McNicholas, K. Raju Jampani, Aaron F. McDaid, T. Brendan Murphy, Larry Banks.

Maintainer: Paul McNicholas <mcnicholas@math.mcmaster.ca>
}
\seealso{
Details, examples, and references are given under \code{\link{pgmmEM}}.
}
