\name{wine}
\alias{wine}
\docType{data}
\title{Italian Wine}
\description{
Data on 27 chemical and physical properties of three types of wine (Barolo, Grignolino, Barbera) from the Piedmont region of Italy. The study did include one further variable but the sulphur measurements were not available.
}
\usage{data(wine)}
\format{
  A data frame with 178 observations and 28 columns. The first column gives the type of wine: (\code{1}) Barolo, (\code{2}) Grignolino, or (\code{3}) Barbera. The other 27 columns contain the variables.
}
\source{
Forina, M., Armanino, C., Castino, M. and Ubigli, M. (1986). Multivariate data analysis as a discriminating method of the origin of wines. \emph{Vitis} \bold{25}, 189--201.
}
\keyword{datasets}
