\name{pgmm}
\alias{pgmm}
\docType{package}
\title{Parsimonious Gaussian Mixture Models}
\description{
An implementation of model-based clustering and model-based classification using parsimonious Gaussian mixture models.
}
\details{
\tabular{ll}{
Package: \tab pgmm\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2015-05-27\cr
License: \tab GPL (>=2)\cr
}
This package contains the function \code{\link{pgmmEM}} plus three data sets.
}
\author{
Paul D. McNicholas [aut, cre], Aisha ElSherbiny [aut], K. Raju Jampani [ctb], Aaron McDaid [aut], Brendan Murphy [aut], Larry Banks [ctb].

Maintainer: Paul D. McNicholas <mcnicholas@math.mcmaster.ca>
}
\seealso{
Details, examples, and references are given under \code{\link{pgmmEM}}.
}
